requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yarl==1.8.2
